Labyrinth visual changes, add sounds

+++ Augmented_climbing_wall/Labyrinth/Labyrinth_config.h
@@ -19,8 +19,8 @@ namespace Labyrinth
 
 
 		sf::Vector2f level1_line1_pos;
-		int level1_line1_length;
-		int level1_line1_angl;
+		float level1_line1_length;
+		float level1_line1_angl;
 		sf::Vector2f level1_startButton;
 		float level1_startButton_radius;
 		sf::Vector2f level1_winButton;
@@ -28,43 +28,43 @@ namespace Labyrinth
 
 
 		sf::Vector2f level2_line1_pos;
-		int level2_line1_length;
-		int level2_line1_angl;
+		float level2_line1_length;
+		float level2_line1_angl;
 		sf::Vector2f level2_line2_pos;
-		int level2_line2_length;
-		int level2_line2_angl;
+		float level2_line2_length;
+		float level2_line2_angl;
 		sf::Vector2f level2_startButton;
 		float level2_startButton_radius;
 		sf::Vector2f level2_winButton;
 		float level2_winButton_radius;
 
 		sf::Vector2f level3_line1_pos;
-		int level3_line1_length;
-		int level3_line1_angl;
+		float level3_line1_length;
+		float level3_line1_angl;
 		sf::Vector2f level3_line2_pos;
-		int level3_line2_length;
-		int level3_line2_angl;
+		float level3_line2_length;
+		float level3_line2_angl;
 		sf::Vector2f level3_startButton;
 		float level3_startButton_radius;
 		sf::Vector2f level3_winButton;
 		float level3_winButton_radius;
-		int level3_angVel;
+		float level3_angVel;
 
 		sf::Vector2f level4_line1_pos;
-		int level4_line1_length;
-		int level4_line1_angl;
+		float level4_line1_length;
+		float level4_line1_angl;
 		sf::Vector2f level4_line2_pos;
-		int level4_line2_length;
-		int level4_line2_angl;
+		float level4_line2_length;
+		float level4_line2_angl;
 		sf::Vector2f level4_line3_pos;
-		int level4_line3_length;
-		int level4_line3_angl;
+		float level4_line3_length;
+		float level4_line3_angl;
 		sf::Vector2f level4_line4_pos;
-		int level4_line4_length;
-		int level4_line4_angl;
+		float level4_line4_length;
+		float level4_line4_angl;
 		sf::Vector2f level4_line5_pos;
-		int level4_line5_length;
-		int level4_line5_angl;
+		float level4_line5_length;
+		float level4_line5_angl;
 		sf::Vector2f level4_startButton;
 		float level4_startButton_radius;
 		sf::Vector2f level4_winButton;
@@ -73,20 +73,20 @@ namespace Labyrinth
 		float level4_button1_radius;
 		sf::Vector2f level4_button2;
 		float level4_button2_radius;
-		int level4_timeToSwithOn_blinkLine;
-		int	level4_timeToSwithOff_blinkLine;
-		int level4_anglVel;
-		int	level4_vel;
+		float level4_timeToSwithOn_blinkLine;
+		float	level4_timeToSwithOff_blinkLine;
+		float level4_anglVel;
+		float	level4_vel;
 
 		sf::Vector2f level5_vel;
-		int	level5_radius;
+		float	level5_radius;
 		sf::Vector2f level5_pos;
-		int	level5_teslaParticls;
+		float	level5_teslaParticls;
 		sf::Vector2f level5_startButton;
 		float level5_startButton_radius;
 		sf::Vector2f level5_winButton;
 		float level5_winButton_radius;
-		int level5_anglVel;
+		float level5_anglVel;
 
 	};
 }
